Helper class for perspective switching, fixed icon
